fix: correct import paths after library reorganization
