+++ src/components/admin/ProjectsModalPaid.svelte
@@ -2,8 +2,8 @@
     import { Modal } from "flowbite-svelte";
 
     import { t } from "../../i18n/store";
-    import Button from "../library/Button.svelte";
-    import TextInput from "../library/TextInput.svelte";
+import Button from "../library/buttons/Button.svelte";
+import TextInput from "../library/inputs/TextInput.svelte";
 
     let {
         open = $bindable(false),
