+++ src/components/library/share/CopyUrl.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import { t } from "../../../i18n/store";
-    import Tooltip from "../../Admin/Tooltip.svelte";
+    import Tooltip from "../../admin/Tooltip.svelte";
     import WebIcon from "../../icons/media/Web.svelte";
 
     interface Props {
