+++ src/pages/[...locale]/admin/projects.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from "../../../layouts/Layout.astro";
-import App from "../../../components/App.svelte";
+import App from "../../../layouts/App.svelte";
 import AdminProjects from "../../../components/admin/Projects.svelte";
 
 const { t, lang, session } = Astro.locals;
